+++ frontend/src/pages/home/ui/HomePage.tsx
@@ -1,5 +1,4 @@
 import { FilterPanel } from "@/widgets/filters/ui/FilterPanel";
-import { SettingsModal } from "@/widgets/settings-modal";
 
 import { ControlsPanel } from "@/pages/home/ui/ControlsPanel";
 import { NavTabs } from "@/pages/home/ui/NavTabs";
@@ -19,7 +18,7 @@ export const HomePage = () => {
 
         <ControlsPanel />
 
-        <SettingsModal />
+        {/* <SettingsModal /> */}
       </nav>
     </div>
   );
